feat(simulation): module de projection des stocks bobines
Phase 3, tache 2.6 du plan issu du CR de reunion PDG. Outil de
projection autonome : il ne lit que l'historique et n'ecrit jamais
rien. Aucune simulation ne peut modifier un stock reel.

Les deux cas d'usage du CR :
- cas 1 : "j'ai N bobines, jusqu'a quand puis-je travailler ?" —
  reponse en clair, du type "1 000 bobines = utilisation jusqu'en
  janvier 2027 au rythme actuel"
- cas 2 : "j'ouvre un site de plus, mon stock encaisse-t-il ?" — sinon
  le nombre de bobines a commander pour tenir l'horizon vise

Parametres tous pilotes par l'URL, donc la simulation est rejouable et
partageable par lien, sans rien stocker : site, fenetre d'historique,
quantite de bobines, films par bobine, consommation (observee ou
saisie), horizon, nombre de nouveaux sites et leur consommation.

Le stock est suivi en films, les questions se posent en bobines : la
conversion passe par films_par_bobine_moyen(), calcule sur le parc reel
plutot que sur une constante. Les valeurs par defaut du formulaire
viennent du stock et de l'historique reels, pour que la page reponde
juste des l'ouverture.

Sortie : verdict en clair, quatre indicateurs, courbe d'evolution en
canvas avec marquage de la rupture, detail de chaque valeur retenue
pour que le resultat puisse etre conteste, et exports Excel et PDF.

Le CR demandait un groupe de navigation dedie ; la page rejoint
RAPPORTS, qui existe deja et accueille les ecrans d'analyse. Creer un
groupe pour une seule page aurait alourdi le menu sans rien clarifier.

── Un defaut de la formule d'origine corrige au passage ──
Le banc de test a revele que la division etait ENTIERE : SUM(quantite)
et l'ecart de dates sont deux entiers, donc 200 films sur 9 jours
donnaient 22 et non 22,22. Surtout, 5 films sur 9 jours donnaient 0 :
consommation declaree nulle alors qu'elle est reelle, et donc aucun
"jours restants" affiche pour cette bobine. L'erreur allait toujours
dans le meme sens — consommation sous-estimee, autonomie surestimee.

Le cast ::numeric corrige les trois fonctions. Consequence assumee :
les jours restants des inventaires deviennent legerement plus courts,
parce que plus justes. Les valeurs conso_quotidienne_moy deja
enregistrees gardent leur chiffre tronque ; seuls les inventaires
ouverts ensuite en beneficient.

Verifie : 21 assertions sur le bloc de calcul reel, execute contre un
PostgreSQL charge du schema complet avec un jeu maitrise (200 films en
9 jours, 2000 films en stock) — autonomie, date d'epuisement, bobines
a commander, ajout d'un site, consommation saisie, consommation nulle
sans division par zero, bornes des parametres, et courbe qui demarre
au stock projete sans jamais devenir negative.

+++ includes/consommation.php
@@ -21,6 +21,21 @@
 //  silencieusement les projections deja affichees dans les inventaires.
 //  Les nouveaux ecrans s'y alignent donc plutot que d'inventer une
 //  seconde definition de la meme grandeur.
+//
+//  ── Un defaut corrige au passage : la division etait entiere ──
+//  SUM(quantite) et l'ecart de dates sont deux entiers : PostgreSQL
+//  faisait donc une division ENTIERE. 200 films sur 9 jours donnaient
+//  22 et non 22,22 ; surtout, 5 films sur 9 jours donnaient 0, soit
+//  une consommation declaree nulle alors qu'elle est reelle — et donc
+//  aucun "jours restants" affiche pour cette bobine.
+//
+//  L'erreur allait toujours dans le meme sens : consommation
+//  sous-estimee, donc autonomie surestimee. Le cast ::numeric la
+//  corrige. Consequence assumee : les jours restants affiches dans les
+//  inventaires deviennent legerement plus courts, parce que plus
+//  justes. Les valeurs conso_quotidienne_moy deja enregistrees dans
+//  inventaire_details_bobines gardent leur ancien chiffre tronque :
+//  seuls les inventaires ouverts apres cette correction en beneficient.
 // ============================================================
 
 /**
@@ -28,7 +43,7 @@
  */
 function conso_moy_bobine(int $bobine_id, int $jours = 30): float {
     return (float) db_fetch_value(
-        "SELECT COALESCE(SUM(quantite) / GREATEST(((NOW())::date - (MIN(date_conso)::date)), 1), 0)
+        "SELECT COALESCE(SUM(quantite)::numeric / GREATEST(((NOW())::date - (MIN(date_conso)::date)), 1), 0)
            FROM consommations_bobines
           WHERE bobine_id = ?
             AND date_conso >= (CURRENT_DATE - (? || ' DAY')::interval)",
@@ -44,7 +59,7 @@ function conso_moy_site(int $site_id = 0, int $jours = 30): float {
     $filtre = $site_id ? "AND site_id = ?" : "";
     $params = $site_id ? [$jours, $site_id] : [$jours];
     return (float) db_fetch_value(
-        "SELECT COALESCE(SUM(quantite) / GREATEST(((NOW())::date - (MIN(date_conso)::date)), 1), 0)
+        "SELECT COALESCE(SUM(quantite)::numeric / GREATEST(((NOW())::date - (MIN(date_conso)::date)), 1), 0)
            FROM consommations_bobines
           WHERE date_conso >= (CURRENT_DATE - (? || ' DAY')::interval)
             $filtre",
@@ -64,7 +79,7 @@ function conso_moy_site(int $site_id = 0, int $jours = 30): float {
 function conso_moy_par_site(int $jours = 30): array {
     $rows = db_fetch_all(
         "SELECT s.id, s.nom,
-                COALESCE(SUM(c.quantite) / GREATEST(((NOW())::date - (MIN(c.date_conso)::date)), 1), 0) AS conso
+                COALESCE(SUM(c.quantite)::numeric / GREATEST(((NOW())::date - (MIN(c.date_conso)::date)), 1), 0) AS conso
            FROM sites s
            LEFT JOIN consommations_bobines c
                   ON c.site_id = s.id
